+++ jdaviz/core/loaders/importers/catalog/catalog.vue
@@ -106,7 +106,7 @@
       :auto.sync="data_label_auto"
       :invalid_msg="data_label_invalid_msg"
       label="Catalog label"
-      api_hint="ldr.importer.label ="
+      api_hint="ldr.importer.data_label ="
       :api_hints_enabled="api_hints_enabled"
       hint="Label to assign to the catalog."
     ></plugin-auto-label>