+++ CHANGES.rst
@@ -190,6 +190,8 @@ Bug Fixes
 
 - Adjust data menu behavior to allow for showing more icons when the viewer is short. [#4352]
 
+- Fix a couple incorrect API hints in the data loaders. [#4399]
+
 Mosviz
 ^^^^^^